homework2.1 finished

+++ homework2.2/src/main/java/Ball.java
@@ -2,22 +2,23 @@ import java.awt.*;
 
 public class Ball extends Sprite {
     private final Color color = new Color(
-            (int)(Math.random() * 255),
-            (int)(Math.random() * 255),
-            (int)(Math.random() * 255)
+            (int) (Math.random() * 255),
+            (int) (Math.random() * 255),
+            (int) (Math.random() * 255)
     );
-    private float vX = (float)(100f + (Math.random() * 200f));
-    private float vY = (float)(100f + (Math.random() * 200f));
+    private float vX = (float) (100f + (Math.random() * 200f));
+    private float vY = (float) (100f + (Math.random() * 200f));
 
     Ball() {
-        halfHeight = 20 + (float)(Math.random() * 50f);
+        halfHeight = 20 + (float) (Math.random() * 50f);
         halfWidth = halfHeight;
     }
-    Ball(int x, int y){
-        halfHeight = 20 + (float)(Math.random() * 50f);
+
+    Ball(float x, float y) {
+        halfHeight = 20 + (float) (Math.random() * 50f);
         halfWidth = halfHeight;
-        this.x=x;
-        this.y=y;
+        this.x = x;
+        this.y = y;
     }
 
     @Override
